+++ lib/design_system/base.rb
@@ -1,12 +1,12 @@
 require_relative 'components/base/breadcrumbs'
-require_relative 'components/form'
+require_relative 'components/base/form'
 require_relative 'components/base/headings'
 
 module DesignSystem
   # This is the base class for design system adapters
   class Base
     include Components::Base::Breadcrumbs
-    include Components::Form
+    include Components::Base::Form
     include Components::Base::Headings
 
     delegate :capture, :content_for, :content_tag, :link_to, :link_to_unless_current, to: :@context
